refactor: extract openProjectConfigPanel handler and fix launch-args JSDoc
Address PR review feedback: extract the debug80.openProjectConfigPanel command
handler from commands.ts into its own src/extension/project-config-panel.ts
module, and replace the placeholder @fileoverview in launch-args.ts with an
explanation of why the file lives at the src/debug/ top level.
